sei messages

+++ H264Decoder/Data/NALU/DecRefPicMarking.h
@@ -9,11 +9,11 @@ struct DecRefPicMarking
 
     bool adaptiveRefPicMarkingModeFlag = false;
 
-    std::uint8_t memoryManagementControlOperation = 0;
-    std::uint8_t differenceOfPicNumsMinus1 = 0;
-    std::uint8_t longTermPicNum = 0;
-    std::uint8_t longTermFrameIdx = 0;
-    std::uint8_t maxLongTermFrameIdxPlus1 = 0;
+    std::uint16_t memoryManagementControlOperation = 0;
+    std::uint16_t differenceOfPicNumsMinus1 = 0;
+    std::uint16_t longTermPicNum = 0;
+    std::uint16_t longTermFrameIdx = 0;
+    std::uint16_t maxLongTermFrameIdxPlus1 = 0;
 
     DecRefPicMarking() = default;
     explicit DecRefPicMarking(NaluTypes nalType, BitstreamReader& reader);
