+++ H264Decoder/Data/NALU/PPSRbsp.h
@@ -8,30 +8,30 @@ struct NALUnit;
 
 struct PPSRbsp
 {
-    std::uint8_t ppsId = 0;
-    std::uint8_t spsId = 0;
+    std::uint16_t ppsId = 0;
+    std::uint16_t spsId = 0;
     bool entropyCodingModeFlag = false;
     bool bottomFieldPicOrderInFramePresentFlag = false;
 
-    std::uint8_t numSliceGroupsMinus1 = 0;
-    std::uint8_t sliceGroupMapType = 0;
-    std::vector<std::uint8_t> runLengthMinus1;
-    std::vector<std::uint8_t> topLeft;
-    std::vector<std::uint8_t> bottomRight;
+    std::uint16_t numSliceGroupsMinus1 = 0;
+    std::uint16_t sliceGroupMapType = 0;
+    std::vector<std::uint16_t> runLengthMinus1;
+    std::vector<std::uint16_t> topLeft;
+    std::vector<std::uint16_t> bottomRight;
     bool sliceGroupChangeDirectionFlag = false;
-    std::uint8_t picSizeInMapUnitsMinus1 = 0;
-    std::uint8_t sliceGroupChangeRateMinus1 = 0;
+    std::uint16_t picSizeInMapUnitsMinus1 = 0;
+    std::uint16_t sliceGroupChangeRateMinus1 = 0;
 
     std::vector<std::uint16_t> sliceGroupId;
 
-    std::uint8_t numRefIdxl0DefaultActiveMinus1 = 0;
-    std::uint8_t numRefIdxl1DefaultActiveMinus1 = 0;
+    std::uint16_t numRefIdxl0DefaultActiveMinus1 = 0;
+    std::uint16_t numRefIdxl1DefaultActiveMinus1 = 0;
 
     bool weightedPredFlag = false;
     std::uint8_t weightedBipredIdc = 0;
-    std::int8_t picInitQPMinus26;
-    std::int8_t picInitQSMinus26;
-    std::int8_t chromaQPIndexOffset;
+    std::int16_t picInitQPMinus26;
+    std::int16_t picInitQSMinus26;
+    std::int16_t chromaQPIndexOffset;
 
     bool deblockingFilterControlPresentFlag = false;
     bool constraintedIntraPredFlag = false;
@@ -43,7 +43,7 @@ struct PPSRbsp
     std::vector<ScalingList> scalingList4x4;
     std::vector<ScalingList> scalingList8x8;
 
-    std::int8_t secondChromaQPIndexOffset;
+    std::int16_t secondChromaQPIndexOffset;
 
 
     RbspTrailingBits rbspTrailingBits;
